+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.16] - 2024-03-22
+
 ### Added
 
 - Support for [Spring Framework](https://docs.spring.io/spring-framework/docs/current/javadoc-api/org/springframework/lang/NonNullFields.html)'s `@NonNullFields` annotation and [JSpecify](https://jspecify.dev/)'s `@NullMarked` annotation. ([Issue 936](https://github.com/jqno/equalsverifier/issues/936))
@@ -1210,7 +1212,8 @@ Please don't use version 1.3; [it's a broken release](https://jqno.nl/post/2013/
 
 You can now use EqualsVerifier!
 
-[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.15.8...HEAD
+[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.16...HEAD
+[3.16]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.15.8...equalsverifier-3.16
 [3.15.8]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.15.7...equalsverifier-3.15.8
 [3.15.7]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.15.6...equalsverifier-3.15.7
 [3.15.6]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.15.5...equalsverifier-3.15.6